cache headers and streams on frame
