+++ src/Microsoft.AspNetCore.Server.Kestrel/ServiceContext.cs
@@ -25,7 +25,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel
             FrameFactory = context.FrameFactory;
             DateHeaderValueManager = context.DateHeaderValueManager;
             ServerOptions = context.ServerOptions;
-            HttpComponentFactory = context.HttpComponentFactory;
         }
 
         public IApplicationLifetime AppLifetime { get; set; }
@@ -39,7 +38,5 @@ namespace Microsoft.AspNetCore.Server.Kestrel
         public DateHeaderValueManager DateHeaderValueManager { get; set; }
 
         public KestrelServerOptions ServerOptions { get; set; }
-
-        internal IHttpComponentFactory HttpComponentFactory { get; set; }
     }
 }
